\name{objfunction-package}
\alias{objfunction-package}
\alias{objfunction}
\docType{package}
\title{
\packageTitle{objfunction}
}
\description{
\packageDescription{objfunction}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{objfunction}
\packageIndices{objfunction}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{objfunction}

Maintainer: \packageMaintainer{objfunction}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
